thesis_to_book/requirements.txt
anthropic>=0.40.0
pdfplumber>=0.11.0
python-docx>=1.1.0
PyYAML>=6.0
tqdm>=4.66.0
